requirements.txt
Flask~=2.0.2
Flask-RESTful==0.3.9
Werkzeug~=2.0.2
flask-mongoengine==1.0.0
mongoengine==0.24.1
PyYAML==6.0
marshmallow==3.15.0
